backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
python-amazon-paapi==6.0.0
requests==2.32.3
pydantic==2.10.3
